pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "faiss-cpu>=1.13.2",
    "flask>=3.1.2",
    "flask-cors>=6.0.2",
    "langchain>=1.2.0",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.6",
    "obd>=0.7.3",
    "openai>=2.14.0",
    "pymupdf>=1.26.7",
    "python-dotenv>=1.2.1",
]
